+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 - Added `ts-node-position-range`.
 - Upgraded `tree-sitter` to 0.6.3. This fixed `ts-type-name-for-id` and `ts-field-name-for-id` crashing on out-of-bounds IDs.
+- Fixed `ts-reset-cursor` always signaling "already mutable borrowed" error.
 
 ## [0.4.0] - 2020-03-01