+++ CHANGELOG.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
-- Added function `ts-node-position-range`.
-- Added function `ts-node-eq`.
+
+## [0.5.0] - 2020-03-17
+- Added functions `ts-node-position-range`, `ts-node-eq`.
+- Added function `tree-sitter-node-at-point`.
 - Added macro `tree-sitter-save-excursion`, which is useful for code formatting operations.
 - Added `tree-sitter-extras` for extra functionalities built on top of `tree-sitter-mode`.
 - Upgraded `tree-sitter` to 0.6.3. This fixed `ts-type-name-for-id` and `ts-field-name-for-id` crashing on out-of-bounds IDs.
@@ -29,7 +31,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.1.0] - 2020-01-27
 Initial release
 
-[Unreleased]: https://github.com/ubolonton/emacs-tree-sitte/compare/0.4.0...HEAD
+[Unreleased]: https://github.com/ubolonton/emacs-tree-sitte/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/ubolonton/emacs-tree-sitte/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/ubolonton/emacs-tree-sitte/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/ubolonton/emacs-tree-sitte/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/ubolonton/emacs-tree-sitte/compare/0.1.0...0.2.0